Changed AgreementEdit component to AgreementRead
